fragmentation first approach and some refactor tasks
