+++ es5/spec/common.index.spec.js
@@ -5,7 +5,7 @@ var assert = require("yeoman-generator").assert;
 var helpers = require("yeoman-generator").test;
 var os = require("os");
 
-describe("troupe", function () {
+describe("troupe:common", function () {
 	var name = undefined;
 
 	before(function (done) {
@@ -16,23 +16,23 @@ describe("troupe", function () {
 		}).on("end", done);
 	});
 
-	it("creates cucumber js files", function () {
+	it("creates common cucumber js files", function () {
 		assert.file(["es6/features/steps/common/jsonWebToken.js", "es6/features/steps/common/language.js", "es6/features/steps/common/request.js", "es6/features/steps/common/values.js", "es6/features/steps/accessToken.steps.js", "es6/features/support/hooks.js", "es6/features/support/world.js"]);
 	});
 
-	it("creates a common step file", function () {
+	it("creates a common steps file", function () {
 		assert.file(["es6/features/steps/common.steps.js"]);
 	});
 
-	it("creates the controllers", function () {
+	it("creates the common controllers", function () {
 		assert.file(["es6/app/controllers/applicationController.js"]);
 	});
 
-	it("creates the managers", function () {
+	it("creates the common manager", function () {
 		assert.file(["es6/app/managers/accountManager.js"]);
 	});
 
-	it("creates misc files", function () {
+	it("creates other common files", function () {
 		assert.file(["es6/app/errors.js", "es6/app/server.js"]);
 	});
 });
