+++ es6/lib/features.index.js
@@ -1,6 +1,6 @@
 var yeoman = require("yeoman-generator");
 
-import inflect from "jargon";
+import {processContext, getPrompts} from "./common.js";
 
 module.exports = yeoman.generators.Base.extend({
 	initializing: function yoInitializing() {
@@ -10,12 +10,7 @@ module.exports = yeoman.generators.Base.extend({
 	prompting: function yoPrompt() {
 		var done = this.async();
 
-		var prompts = [{
-			type: "input",
-			name: "name",
-			message: "What is the model name? (use camel case please)",
-			default: "myModel"
-		}];
+		var prompts = getPrompts();
 
 		this.prompt(prompts, function (props) {
 			this.props = props;
@@ -25,12 +20,22 @@ module.exports = yeoman.generators.Base.extend({
 	},
 
 	writing: function yoWriting() {
-		const context = {
-			name: this.props.name,
-			Name: inflect(this.props.name).pascal.toString(),
-			names: inflect(this.props.name).plural.toString(),
-			_name: inflect(this.props.name).snake.toString()
-		};
+		const context = processContext(this.props);
+
+		//copy feature steps
+		["_model.show.steps.js",
+		"_model.create.steps.js",
+		"_model.update.steps.js",
+		"_model.delete.steps.js",
+		"_model.list.steps.js"]
+		.forEach((templatePath) => {
+			let newName = templatePath.replace("_model", `${context.name}`);
+			this.fs.copyTpl(
+				this.templatePath(`es6/features/steps/${templatePath}`),
+				this.destinationPath(`es6/features/steps/${context.name}/${newName}`),
+				context
+			);
+		}, this);
 
 		//copy features
 		["_model.show.feature",
@@ -41,7 +46,7 @@ module.exports = yeoman.generators.Base.extend({
 		.forEach((templatePath) => {
 			let newName = templatePath.replace("_model", `${context.name}`);
 			this.fs.copyTpl(
-				this.templatePath("es6/features/" + templatePath),
+				this.templatePath(`features/${templatePath}`),
 				this.destinationPath(`features/${context.name}/${newName}`),
 				context
 			);
@@ -52,7 +57,7 @@ module.exports = yeoman.generators.Base.extend({
 		.forEach((templatePath) => {
 			let newName = templatePath.replace("_", "");
 			this.fs.copyTpl(
-				this.templatePath("es6/features/" + templatePath),
+				this.templatePath(`features/${templatePath}`),
 				this.destinationPath(`features/${newName}`),
 				context
 			);
