+++ es6/spec/app.index.spec.js
@@ -18,16 +18,6 @@ describe("troupe", () => {
 			.on("end", done);
 	});
 
-	it("creates cucumber js files", () => {
-		assert.file([
-			`es6/features/steps/${name}/${name}.show.steps.js`,
-			`es6/features/steps/${name}/${name}.create.steps.js`,
-			`es6/features/steps/${name}/${name}.update.steps.js`,
-			`es6/features/steps/${name}/${name}.delete.steps.js`,
-			`es6/features/steps/${name}/${name}.list.steps.js`,
-		]);
-	});
-
 	it("creates the model and his spec", () => {
 		assert.file([
 			`es6/app/models/${name}.js`,
@@ -55,10 +45,6 @@ describe("troupe", () => {
 	});
 
 	describe("(about the content of every feature)", () => {
-		it("should have some this.database = on it", () => {
-			assert.fileContent([[`es6/features/steps/${name}/${name}.show.steps.js`, "this.database"]]);
-		});
-
 		describe("(controller content)", () => {
 				it("should have filters for validation", () => {
 						assert.fileContent([
