+++ es6/spec/features.index.spec.js
@@ -18,6 +18,16 @@ describe("troupe:features", () => {
 			.on("end", done);
 	});
 
+	it("creates cucumber js files", () => {
+		assert.file([
+			`es6/features/steps/${name}/${name}.show.steps.js`,
+			`es6/features/steps/${name}/${name}.create.steps.js`,
+			`es6/features/steps/${name}/${name}.update.steps.js`,
+			`es6/features/steps/${name}/${name}.delete.steps.js`,
+			`es6/features/steps/${name}/${name}.list.steps.js`
+		]);
+	});
+
 	it("creates feature files", () => {
 		assert.file([
 			`features/${name}/${name}.show.feature`,
@@ -27,5 +37,11 @@ describe("troupe:features", () => {
 			`features/${name}/${name}.list.feature`,
 			`features/accessToken.feature`
 		]);
+	});
+
+	describe("(about the content of every feature)", () => {
+		it("should have some this.database = on it", () => {
+			assert.fileContent([[`es6/features/steps/${name}/${name}.show.steps.js`, "this.database"]]);
+		});
 	});
 });
