+++ generators/features/templates/es6/features/steps/_model.delete.steps.js
