+++ generators/features/templates/features/_model.delete.feature
