+++ generators/features/templates/features/_model.show.feature
